Add files via upload
This game is called Hangman. It is a classic word guessing game where one player thinks of a word and the other tries to guess it by suggesting letters within a certain number of guesses. The player guesses letters one at a time, and if the guessed letter is in the word, it is revealed in its correct positions. If the letter is not in the word, the player loses a guess. The game ends when the player either guesses all the letters in the word or runs out of guesses. The user has 15 trials.